Rudnyk/requirements.txt
arrow==0.12.0
binaryornot==0.4.4
cassandra-driver==3.12.0
certifi==2017.11.5
chardet==3.0.4
click==6.7
flake8==3.5.0
Flask==0.12.2
future==0.16.0
idna==2.6
itsdangerous==0.24
Jinja2==2.10
jinja2-time==0.2.0
MarkupSafe==1.0
mccabe==0.6.1
poyo==0.4.1
pycodestyle==2.3.1
pyflakes==1.6.0
PyQt5==5.9.2
python-dateutil==2.6.1
requests==2.18.4
sip==4.19.6
six==1.11.0
urllib3==1.22
Werkzeug==0.13
whichcraft==0.4.1
